+++ tests/Unit/TokenTest.php
@@ -45,7 +45,7 @@ class TokenTest extends TestCase
     public function it_defaults_sending_token_as_sms()
     {
         $url = $this->api."sms/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => null, 'actionMessage' => null]];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => null, 'actionMessage' => null]];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->send($this->validAuthyId));
@@ -55,7 +55,7 @@ class TokenTest extends TestCase
     public function it_sends_token_as_phone_call()
     {
         $url = $this->api."call/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => null, 'actionMessage' => null]];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => null, 'actionMessage' => null]];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->send($this->validAuthyId, 'call'));
@@ -65,7 +65,7 @@ class TokenTest extends TestCase
     public function it_enforces_sending_token_over_cellphone_network()
     {
         $url = $this->api."sms/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => true, 'action' => null, 'actionMessage' => null]];
+        $params = $this->params + ['query' => ['force' => 'true', 'action' => null, 'actionMessage' => null]];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->send($this->validAuthyId, 'sms', true));
@@ -75,7 +75,7 @@ class TokenTest extends TestCase
     public function it_sets_action_and_action_message()
     {
         $url = $this->api."sms/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => 'login', 'actionMessage' => 'Login Code']];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => 'login', 'actionMessage' => 'Login Code']];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->send($this->validAuthyId, 'sms', false, 'login', 'Login Code'));
@@ -85,7 +85,7 @@ class TokenTest extends TestCase
     public function it_verifies_token()
     {
         $url = $this->api."verify/{$this->validToken}/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => null]];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => null]];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->verify($this->validToken, $this->validAuthyId));
@@ -95,7 +95,7 @@ class TokenTest extends TestCase
     public function it_enforces_token_verification_regardless_registration_completeness()
     {
         $url = $this->api."verify/{$this->validToken}/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => true, 'action' => null]];
+        $params = $this->params + ['query' => ['force' => 'true', 'action' => null]];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->verify($this->validToken, $this->validAuthyId, true));
@@ -105,7 +105,7 @@ class TokenTest extends TestCase
     public function it_verifies_token_action()
     {
         $url = $this->api."verify/{$this->validToken}/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => 'login']];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => 'login']];
 
         $this->http->shouldReceive('get')->once()->with($url, $params)->andReturn($this->httpResponse);
         $this->assertEquals($this->authyResponse, $this->authyToken->verify($this->validToken, $this->validAuthyId, false, 'login'));
@@ -115,7 +115,7 @@ class TokenTest extends TestCase
     public function it_returns_errors_array_when_request_fails_and_errors_exists()
     {
         $url = $this->api."sms/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => null, 'actionMessage' => null]];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => null, 'actionMessage' => null]];
         $httpResponse = new HttpResponse(200, [], json_encode(['success' => false, 'errors' => ['foo' => 'bar']]));
         $authyResponse = new AuthyResponse($httpResponse);
 
@@ -129,7 +129,7 @@ class TokenTest extends TestCase
     public function it_returns_empty_errors_array_when_request_success_even_when_errors_exists()
     {
         $url = $this->api."sms/{$this->validAuthyId}";
-        $params = $this->params + ['query' => ['force' => false, 'action' => null, 'actionMessage' => null]];
+        $params = $this->params + ['query' => ['force' => 'false', 'action' => null, 'actionMessage' => null]];
         $httpResponse = new HttpResponse(200, [], json_encode(['success' => true, 'errors' => ['foo' => 'bar']]));
         $authyResponse = new AuthyResponse($httpResponse);